The 6 tasks of Homework 1 in one commit.
Task 1: Compile the Java code listed in Logic.java and Main.java (Done: https://imgur.com/G9MNiVI)

Task 2: Find a greatest N which complies with following conditions
a_n < epsilon;
a_n = 1/(N+1)^2

Task 3: Solve F(x) for range [a; b] with the step of h.
F(x) = tan(2x)-3
Show the results as a spreadsheet which has x as a first column, and F(x) as the second one.

Task 4: The set of numbers (a_1, a_2, ... a_N) is specified.
Find max(a_1+a_N, a_2+a_N-1, ... a_N/2+a_N/2+1)

Task 5: Get a matrix which has ones in the main and counter diagonals and zeroes in the rest of the entries.

Task 6: Make classes for a notebook entry and a notebook (the entries should be stored in an array). Implement following entries:
* Add an entry (Notebook.addEntry())
* Remove an entry (Notebook.removeEntry())
* Edit an entry (Notebook.editEntry())
* Show all entries (Notebook.getAllTitles, Notebook.getAllStrings)
Make JavaDoc for them.